hnsw/pyproject.toml
[project]
name = "hnsw"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.12.0",
    "numba>=0.61.2",
    "numpy>=2.2.6",
    "polars>=1.32.3",
]
